be_fastapi/requirements.txt
uvicorn
fastapi
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
pydantic-settings